Fixed one-pixel-sized specles in the diffusion model due to bilinear upsampling in the last decoder layer.
In your latest model where you write:

self.up = nn.Upsample(scale_factor=2, mode='bilinear', align_corners=False)

doesn't that make it stop caring about individual pixels -- which results in a bunch
of one-pixel-sized speckley noise when used as a diffusion model?

We had a similar problem with a model that had:
 u_final = F.interpolate(u1, size=x.shape[-2:], mode='bilinear', align_corners=False)
and we replaced it with an additional d0 = ... layer similar to the d1, d2, and d3 layer?